GNG-570: Move new component selection fields to AddNew tab
